backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-MySQL==1.5.2
python-dotenv==1.0.0